master Delete:470 execute proxy command only if instance is up

+++ app/src/Commands/Instances/DeleteCommand.php
@@ -139,10 +139,12 @@ class DeleteCommand extends Command
         $this->docker->removeContainer($instance . "_" . $suffix);
 
         $this->docker->executeCommand(self::SALT_MAIN, "doil_saltmain", "salt-key", "-d", "$instance.$suffix", "-y", "-q");
-        $this->docker->executeDockerCommand(
-            "doil_proxy",
-            "rm -f /etc/nginx/conf.d/sites/$instance.conf &&  /root/generate_index_html.sh"
-        );
+        if ($this->docker->isInstanceUp($path)) {
+            $this->docker->executeDockerCommand(
+                "doil_proxy",
+                "rm -f /etc/nginx/conf.d/sites/$instance.conf &&  /root/generate_index_html.sh"
+            );
+        }
 
         if ($this->docker->hasVolume($instance)) {
             $this->docker->removeVolume($instance);
